feat(tcp-proxy.c): add tcp-proxy.c file to implement a basic TCP proxy server
The tcp-proxy.c file is added to the project. This file contains the implementation of a basic TCP proxy server using the libuv library. The proxy server listens on CLIENT_PORT (7999) and forwards the incoming connections to the REMOTE_SERVER_IP (192.168.0.188) and REMOTE_SERVER_PORT (443). The proxy_t struct is defined to hold the necessary information for each connection, including the client and remote sockets, a write request, and a buffer. The on_close function is used to free the allocated memory for the socket when the connection is closed. The on_remote_read function is responsible for reading from the remote system and forwarding the data to the client. If an error occurs or the connection is closed, the appropriate cleanup is performed.